fix response in ReceiveCaptureBufferInfo

+++ src/core/hle/service/apt/apt.cpp
@@ -567,7 +567,7 @@ void Module::Interface::ReceiveCaptureBufferInfo(Kernel::HLERequestContext& ctx)
 
     IPC::RequestBuilder rb = rp.MakeBuilder(2, 2);
     rb.Push(RESULT_SUCCESS);
-    rb.Push(apt->screen_capture_buffer.size());
+    rb.Push(static_cast<u32>(apt->screen_capture_buffer.size()));
     rb.PushStaticBuffer(std::move(apt->screen_capture_buffer), 0);
 }
